style: move FlowUtils to a package without underscore
